consertando erros

+++ Enemy.java
@@ -23,7 +23,7 @@ public abstract class Enemy extends Game_Explodable{
         explode(500);
     }
 
-    protected boolean hasExploded(int currentTime){ // avalia se o inimigo foi explodido
+    protected boolean hasExploded(long currentTime){ // avalia se o inimigo foi explodido
         if(state = EXPLODING && currentTime > explosionEnd) return true;
         else return false;
     }
